QTzTimeZonePrivate: replace an inefficient QList with QVector (II)
The implementation-private QTzTransition type is larger than void*,
so holding them in QLists is horribly inefficient.

Fix by declaring it as a primitive types (it just contains
various integers) and holding it in QVector instead.

Also optimize parseTzTransitions() by preallocating the expected
number of transitions and streaming directly into the container,
resizing to the number of successful read items on error.

Saves roughly 0.5K in text size, too.

Change-Id: Iadec7a7b0721893e477e1778c9fb54afd6e49544
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>