QTzTimeZonePrivate: replace an inefficient QList with QVector (III)
The implementation-private QTzType type is larger than void*,
so holding them in QLists is horribly inefficient.

Fix by  holding it in QVector instead (it was already marked
as a primitive type before).

Text size grows by ca. 0.5K, but of course we got rid of all
those pesky heap allocations.

Change-Id: I3b70ed36fa9947b695ffc87c6f6199daa13cb7cd
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>