Windows: Remove "experimental" sticker for DirectWrite engine
This engine has been regarded "experimental" for a long time, but
it is in active use by some applications and rather stable.

While if may be replaced by FreeType at some point, if support
for Microsoft's colored fonts is introduced there, at this point
it really doesn't make sense to call it "experimental" anymore.

Change-Id: I038d71b121e90bd7b4dbc6d0afa7b14c9b824b08
Reviewed-by: Andy Shaw <andy.shaw@digia.com>