Added new private API for Android and onPause/onResume
On Android the foreground activity can get onPause/onResume
calls. In Nfc code in android we need to know if we are paused
or resumed. And we need to make sure to call disableForegroundDispatch
inside the onPause call and in the main Ui thread. The current
implementiton of applicationStateChanged was not sufficient to
acomplish that.

This patch adds a way to receive those onPause/onResume calls in Qt
with proper timing.

Change-Id: I3a8cec093fc02ec42cc8677dfe2d0d4f8a227f8b
Reviewed-by: BogDan Vatra <bogdan@kde.org>