Fix compilation with ICC: failure to parse placement of Q_DECL_OVERRIDE
This is the right way:
  void (*getProcAddress(const QByteArray &procName)) () Q_DECL_OVERRIDE;

However, ICC and GCC do not get it. They expect it in the form:
  void (*getProcAddress(const QByteArray &procName) Q_DECL_OVERRIDE) ();

ICC complains with an error, so this needs a workaround. GCC ignores the
incorrectly-placed "override", so it just looks to GCC like no "override"
was added.

See also: https://groups.google.com/a/isocpp.org/forum/#!topic/std-discussion/IjVB8CWiS5A
Change-Id: I1a800c709d3543699131ffff13c2ffc9cee46245
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>