Fix android menubar's menu not opening at the first press.
In case of selected item in submenu, call aboutToHide() on still
visible menus. This will allow quickquickcontrols to act
poperly when menus are closed.

This patch is a resubmission of
dbd6d158e3002689c4c367709d68744465ffd9ec which was merged to dev
although it should have targeted 5.5.

Task-number: QTBUG-42188
Change-Id: I868cc6fc8403f0586f35a1734e0cb026982f6063
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>