Restore binary compatibility in versioned OpenGL function classes
This is a partial revert of commit
18aae36a90c0753f1b1e615ba8437d8ebd1bd2fb that restores the removed 
*Backend classes and restores the ordering of members in other
*Backend classes. This restores binary compatibility for QtGui.

tst_bic now reports only that some of the *Backend classes have
changed size. This is a false positive as all such cases have added
new members at the end of the respective classes. Offsets to
pre-existing members are still consistent with earlier versions.

Change-Id: Ia7f651870c2dea7f1a8d3611e667151387506dc8
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>