Fix underline for centered text
The position of the text itself is always floored, since the glyph
contains the subpixel antialiasing. Since no antialiasing is applied
to the underline, its position also has to be floored, otherwise
it will sometimes appear one pixel to the right of the text it's
underlining.

[ChangeLog][Text] Fixed position of underline on centered text when
the text layout's width is an even number.

Task-number: QTBUG-39286
Change-Id: Ib1704ac5b09289f5b3f5a6580443f234d6d868fc
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>