avoid sending QPinchGestures based on unreasonable pinch deltas
On some devices there can be spurious events indicating that the two
touchpoints are very close together, so that the ratio of present
line length to previous line length can be very large, and the angle
delta can be random.  But in the normal scenario, there will be a lot
of events, in which the touch point movements are small.  So if the
line length ratio is unreasonable, it's safe to ignore one event and
wait for the next.

Task-number: QTBUG-44665
Change-Id: Ibb7fe560b6a3f7db72d73aad3468c61f24a95d13
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>